+++ services/web/client/src/components/dialogs/SessionLobbyDialogDynamic.tsx
@@ -1,0 +1,12 @@
+"use client";
+
+import dynamic from "next/dynamic";
+
+const SessionLobbyDialog = dynamic(
+  () => import("@/components/dialogs/SessionLobbyDialog").then((m) => ({ default: m.SessionLobbyDialog })),
+  { ssr: false },
+);
+
+export default function SessionLobbyDialogDynamic() {
+  return <SessionLobbyDialog />;
+}
